Editor bug fixed & code refactoried
