+++ src/App.js
@@ -1,9 +1,4 @@
 import MainApp from "./components/MainApp";
-// import List from "./components/List";
-// import AddMemo from "./components/AddMemo";
-// import Editor from "./components/Editor";
-// import { useState } from "react";
-
 import "antd/dist/antd.css";
 const App = () => {
